Recomendações através de tags

+++ src/main/java/jonatan/andrei/domain/UserActionUpdateType.java
@@ -3,13 +3,15 @@ package jonatan.andrei.domain;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public enum UserActionUpdateType {
 
-    INCREASE(1),
-    DECREASE(-1);
+    INCREASE(BigDecimal.valueOf(1)),
+    DECREASE(BigDecimal.valueOf(-1));
 
-    private int value;
+    private BigDecimal value;
 
 }
